% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstanr_utils.R
\name{compile_stan_model}
\alias{compile_stan_model}
\title{compile a stan model}
\usage{
compile_stan_model(
  stan_file,
  compile = TRUE,
  dir = file.path("stanmodels", "compiled_stan_models"),
  ...
)
}
\arguments{
\item{stan_file}{(string) The path to a \code{.stan} file containing a Stan
program. The helper function \code{\link[cmdstanr:write_stan_file]{write_stan_file()}} is provided for cases when
it is more convenient to specify the Stan program as a string. If
\code{stan_file} is not specified then \code{exe_file} must be specified.}

\item{compile}{(logical) Do compilation? The default is \code{TRUE}. If \code{FALSE}
compilation can be done later via the \code{\link[cmdstanr:model-method-compile]{$compile()}}
method.}

\item{dir}{directory for compiled executables}

\item{...}{Optionally, additional arguments to pass to the
\code{\link[cmdstanr:model-method-compile]{$compile()}} method if \code{compile=TRUE}.}
}
\description{
compile a stan model
}
\details{
This is a helper for this project so you don't have to adjust any directories.
It's not recommended for use outside of this project, just use cmdstanr::cmdstan_model directly.
}
