% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skew_logistic.R
\name{dELGW}
\alias{dELGW}
\alias{pELGW}
\alias{qELGW}
\alias{rELGW}
\title{The exponential logistic generalized Weibull distribution}
\usage{
dELGW(x, location, scale, shape, log = FALSE)

pELGW(q, location, scale, shape, log = FALSE)

qELGW(p, location, scale, shape, log = FALSE)

rELGW(n, location, scale, shape)
}
\description{
The exponential logistic generalized Weibull distribution
}
\details{
see Aljarrah, M.A., Famoye, F. & Lee, C. Generalized logistic distribution and its regression model. J Stat Distrib App 7, 7 (2020). https://doi.org/10.1186/s40488-020-00107-8
}
