% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstanr_utils.R
\name{save_or_run_model}
\alias{save_or_run_model}
\title{run a stan model}
\usage{
save_or_run_model(expr, path, force)
}
\arguments{
\item{expr}{some R expression}

\item{path}{path to saved results}

\item{force}{if TRUE, ignore any saved results the}
}
\description{
run a stan model
}
\details{
if path exists and force is FALSE it loads the results with \code{readRDS}. Otherwise, evaluate \code{expr}.
}
